Changed the offset method
Previously, the span offsets were found by using offsetTop. This is property is relative to the parent of the element, which doesn't give the offset relative to the top of the document. I changed this to $(element).offset().top, which gives the offset relative to the top of the document.